endpoint: move regeneration functions to `regenerate.go`
Signed-off by: Ian Vernon <ian@cilium.io>
